% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Sample meta data: metadata}
\format{
A data frame with A rows and B columns:
\describe{
  \item{Meta data}{This is the simulated meta data for demonstration purposes, which includes Abx, age and study.}
}
}
\source{
Simulated data
}
\usage{
metadata
}
\description{
Meta data for demonstration purposes.
}
\keyword{datasets}
