% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{otu.tab}
\alias{otu.tab}
\title{Sample otu table: otu.tab}
\format{
A Matrix with 30 rows and 20 columns:
\describe{
  \item{rows}{Each row is a sample.}
  \item{columns}{Each column is the abundance of Taxon.}
}
}
\source{
Simulated data
}
\usage{
otu.tab
}
\description{
A sample otu table for demonstration purposes.
}
\keyword{datasets}
